fix: aba Notas Fiscais quebrava ao renderizar valor da nota
valor_total é numeric no Postgres e volta como string pro node-pg. A
página chama .toFixed(2) em cima disso (linha a linha e no somatório),
que estoura porque string não tem esse método — derrubava a aba
inteira. Cast pra ::float na query resolve nos dois lugares.

+++ frontend/app/api/fiscal/notas/route.ts
@@ -25,7 +25,7 @@ export async function GET(req: Request) {
     const { rows } = await pool.query(`
       SELECT
         fn.id, fn.status, fn.numero, fn.serie, fn.chave_acesso AS "chaveAcesso",
-        fn.motivo_rejeicao AS "motivoRejeicao", fn.valor_total AS "valorTotal",
+        fn.motivo_rejeicao AS "motivoRejeicao", fn.valor_total::float AS "valorTotal",
         fn.ambiente, fn.criado_em AS "criadoEm", fn.autorizado_em AS "autorizadoEm",
         fn.enviado_email_em AS "enviadoEmailEm", fn.enviado_whatsapp_em AS "enviadoWhatsappEm",
         string_agg(o.number, ', ' ORDER BY o.number) AS "orderNumbers",
